Реализация загрузки с файла и объектов.

+++ TileAnim.h
@@ -13,7 +13,7 @@ protected:
 public:
 
 	TileAnim();
-	TileAnim(std::string Name, short int inter, short int texture_size_w, short int texture_size_h, char ass, short int anim_f, short int anim_q, float time);
+	TileAnim(std::string Name, short int inter, float texture_size_w, float texture_size_h, char ass, short int anim_f, short int anim_q, float time);
 	void tile_animation();
 
 };
